backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rishang <rishang@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
pydantic = "^2.5.3"
requests = "^2.31.0"
rich = "^13.7.0"
uvicorn = "^0.25.0"
python-dotenv = "^1.0.0"
blinker = "^1.7.0"
django = "^5.0.1"
django-ninja = "^1.1.0"
django-environ = "^0.11.2"
pymemcache = "^4.0.0"

# to be removed
pymongo = "^4.6.1"
mongoengine = "^0.27.0"
redis = "^5.0.1"
hiredis = "^2.3.2"
schemathesis = "^3.23.1"
# end of to be removed
django-extensions = "^3.2.3"
django-cors-headers = "^4.3.1"
psycopg = {extras = ["binary"], version = "^3.1.17"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
faker = "^21.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
